Done — I’m not sure how, but I just finished the task :)

+++ firmware/include/digital_manager.h
@@ -5,6 +5,6 @@
 
 void setup_digital_manager();
 void digital_manager(void *pvParameters);
-void fan_set_speed(uint8_t speed);
+void fan_set_speed(int16_t speed);
 
 #endif // __DIGITAL_MANAGER_H__
