+++ firmware/include/global.h
@@ -62,7 +62,7 @@ extern uint8_t ws2812_brightness;   // 0..255, 0 = off
 extern boolean is_relay_on;
 
 extern boolean is_mini_fan_on;
-extern uint8_t fan_speed;           // 0..255, 0 = off (PWM duty)
+extern int16_t fan_speed;           // 0..4095, 0 = off (PWM duty)
 
 extern float glob_inference_result;
